feat: add product layout and detail types, update README
