+++ src/types/product.d.ts
@@ -1,0 +1,16 @@
+interface ProductDetail {
+  title: string;
+  price: number;
+  discountPercentage: number;
+  rating: number;
+  stock: number;
+  brand: string;
+  category: string;
+  thumbnail: string;
+  images: string[];
+  tags: string[];
+  warrantyInformation?: string;
+  shippingInformation?: string;
+  availabilityStatus?: string;
+  minimumOrderQuantity?: number;
+}
